requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.2
django-allauth==0.35.0
django-filter==1.1.0
djangorestframework==3.7.7
elasticsearch==6.1.1
elasticsearch-dsl==6.1.0
idna==2.6
ipaddress==1.0.19
Markdown==2.6.11
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
